ci: move release code to separate file
